refactor(ui): remove stale workflow monitor remnants
